Separate Data Queue Universe Provider implementation file
